% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TfuncViz.R
\name{calculate_tfunc_scores}
\alias{calculate_tfunc_scores}
\title{Calculate T cell functionality scores}
\usage{
calculate_tfunc_scores(
  seurat_obj,
  gene_sets = NULL,
  cluster_col = "seurat_clusters",
  group_col = "group",
  ctrl = 100
)
}
\arguments{
\item{seurat_obj}{A Seurat object}

\item{gene_sets}{A list of gene sets for scoring}

\item{cluster_col}{Column name for clusters in metadata}

\item{group_col}{Column name for group in metadata}

\item{ctrl}{Number of control genes for AddModuleScore}
}
\value{
A Seurat object with scores and a list of analysis results
}
\description{
Calculate T cell functionality scores
}
