% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TfuncViz.R
\name{plot_tfunc_bubble_diff}
\alias{plot_tfunc_bubble_diff}
\title{Create a polar bubble plot of functionality differences}
\usage{
plot_tfunc_bubble_diff(tfunc_results)
}
\arguments{
\item{tfunc_results}{Results from calculate_tfunc_scores}
}
\value{
A ggplot object
}
\description{
Create a polar bubble plot of functionality differences
}
