% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TfuncViz.R
\name{plot_tfunc_facet}
\alias{plot_tfunc_facet}
\title{Create faceted group comparison plot}
\usage{
plot_tfunc_facet(
  tfunc_results,
  title = "Comparison of Functional State Distribution Between Groups"
)
}
\arguments{
\item{tfunc_results}{Results from calculate_tfunc_scores}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
Create faceted group comparison plot
}
