% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TfuncViz.R
\name{plot_tfunc_heatmap}
\alias{plot_tfunc_heatmap}
\title{Create a heatmap of T cell functionality scores}
\usage{
plot_tfunc_heatmap(tfunc_results, filename = NULL, width = 12, height = 8)
}
\arguments{
\item{tfunc_results}{Results from calculate_tfunc_scores}

\item{filename}{Optional filename to save PDF}

\item{width}{PDF width}

\item{height}{PDF height}
}
\value{
A ComplexHeatmap object
}
\description{
Create a heatmap of T cell functionality scores
}
