% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TfuncViz.R
\name{plot_tfunc_parallel}
\alias{plot_tfunc_parallel}
\title{Create a parallel coordinates plot of T cell functionality}
\usage{
plot_tfunc_parallel(tfunc_results)
}
\arguments{
\item{tfunc_results}{Results from calculate_tfunc_scores}
}
\value{
A ggplot object
}
\description{
Create a parallel coordinates plot of T cell functionality
}
