% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TfuncViz.R
\name{plot_tfunc_vector}
\alias{plot_tfunc_vector}
\title{Create vector plot showing direction of change}
\usage{
plot_tfunc_vector(
  tfunc_results,
  title = "Disease-Health Group Functional State Change Vectors",
  subtitle =
    "Arrows indicate the direction of functional state changes in the disease group relative to the healthy group"
)
}
\arguments{
\item{tfunc_results}{Results from calculate_tfunc_scores}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}
}
\value{
A ggplot object
}
\description{
Create vector plot showing direction of change
}
